Splitting up types into v1 and v2 modules
