+++ src/generators/correct/mod.rs
@@ -1,0 +1,4 @@
+//! Proptest strategies for most common types
+pub mod primitive;
+pub mod v1;
+pub mod v2;
