+++ src/lib.rs
@@ -1,11 +1,5 @@
-pub mod address;
-pub mod crypto;
-pub mod datum;
 pub(crate) mod feature_traits;
 pub mod generators;
-pub mod interval;
 pub mod plutus_data;
-pub mod redeemer;
-pub mod script;
-pub mod transaction;
-pub mod value;
+pub mod v1;
+pub mod v2;
