+++ src/v1/crypto.rs
@@ -1,5 +1,5 @@
 //! Types for cryptograhic primitives, and other lower level building blocks
-use crate::plutus_data::{PlutusData, PlutusDataError, PlutusType, IsPlutusData};
+use crate::plutus_data::{IsPlutusData, PlutusData, PlutusDataError, PlutusType};
 #[cfg(feature = "lbf")]
 use data_encoding::HEXLOWER;
 #[cfg(feature = "lbf")]
