+++ src/v1/value.rs
@@ -1,9 +1,9 @@
 //! Types related to Cardano values, such as Ada and native tokens.
-use crate::crypto::LedgerBytes;
 use crate::plutus_data::{
-    verify_constr_fields, PlutusData, PlutusDataError, PlutusType, IsPlutusData,
+    verify_constr_fields, IsPlutusData, PlutusData, PlutusDataError, PlutusType,
 };
-use crate::script::{MintingPolicyHash, ScriptHash};
+use crate::v1::crypto::LedgerBytes;
+use crate::v1::script::{MintingPolicyHash, ScriptHash};
 #[cfg(feature = "lbf")]
 use lbr_prelude::json::{Error, Json, JsonType};
 use num_bigint::BigInt;
