+++ src/v2/mod.rs
@@ -1,0 +1,13 @@
+//! Plutus types and utilities for Plutus V2
+//!
+//! Types and utilites unchanged in the new version are re-exported from the v1 module.
+pub mod datum;
+pub mod transaction;
+
+// Inherited from v1
+pub use crate::v1::address;
+pub use crate::v1::crypto;
+pub use crate::v1::interval;
+pub use crate::v1::redeemer;
+pub use crate::v1::script;
+pub use crate::v1::value;
